refactor: enum CalculatorSign을 extension String으로 변경
StringType.swift 파일 추가
